requirements.txt
langchain
langchain-astradb
langchain-huggingface
langchain-community
datasets
pypdf 
python-dotenv
pandas
flask
prometheus_client
